qutebrowser: add support for per domain settings
Add `programs.qutebrowser.perDomainSettings` which let's one to set
configuration options for specific URLs [1].

It option doesn't check if the options passed to it are valid, it
translates the config to python code to be written on the file as is.
Mimicking the behaviour of `programs.qutebrowser.settings`.

Added a new test case `test-qutebrowser-url-settings` for testing the
implementation.

[1]: https://github.com/qutebrowser/qutebrowser/blob/bb7bbb6ead5b7412fa5828249d56e51c1b35cb5a/doc/help/configuring.asciidoc#per-domain-settings